AP-7074: Add TrainingData#getTargetDimension
In classification the number of classes can be seen as the dimension of the target variable.
The addition of this method makes the class ClassificationTrainingData superfluous since it only specifies only a single method that is equivalent to the new method.
Could also enable us to perform regression with multiple target variables.