AP-7057: FileStoreCells in loops and remote execution
With remote execution file stores get initialized in
mimicRemotePreExecute(). But this method is called on the nodes of a
wfm in a random order (actually the order of creation). For file stores
in loops the loop start node needs to create a special FileStoreHandler
which is used by any node inside the loop. But when we try to initialize
file stores on a node inside the loop before the loop start this special
FileStoreHandler isn't created yet.

The solution is to check if mimicRemotePreExecute() was already called
on the loop start node and if not call it directly.
We also need to use the FlowLoopContext of the loop start node because
if the current node isn't configured it hasn't the most current
FlowLoopContext.

Another approach would be to make sure mimicRemotePreExecute gets called
in the right order from the WorkflowManager but this isn't that trivial
and there is still the problem that unconfigured nodes doesn't have the
most current FlowLoopContext.